fix: repair compilation errors in Core.Accounting.Cache, HR.Types, Integration.Health, Production.Project

+++ src/HR/Types.hs
@@ -79,7 +79,7 @@ mkSalarySummary charges = SalarySummary
   }
 
 -- | Make salary charge (stub)
-mkSalaryCharge :: Int64 -> Double -> SalaryCharge
+mkSalaryCharge :: Double -> Int64 -> SalaryCharge
 mkSalaryCharge amt id = SalaryCharge
   { salaryChargeId = id
   , salaryChargeAmount = amt
